dont try to load from cache when initially caching

+++ src/library/services/graphics/font/Letter.h
@@ -18,8 +18,8 @@ struct Letter {
     } offset;
 
     struct Advance {
-        int x;
-        int y;
+        long x;
+        long y;
     } advance;
 };
